requirements.txt
tweepy==4.15.0
requests==2.32.3
python-dotenv==1.0.1
